refactor & clean up server
